requirements.txt
nltk
pydub
Pillow
pytesseract
SpeechRecognition